migrate instance logic to more intuitive location
