+++ pkg/vault/client.go
@@ -7,130 +7,10 @@ import (
 	"fmt"
 	"os"
 	"strings"
-	"sync"
 
-	"github.com/app-sre/vault-manager/pkg/utils"
 	"github.com/hashicorp/vault/api"
 	log "github.com/sirupsen/logrus"
 )
-
-type AuthBundle struct {
-	SecretEngine string
-	VaultSecrets []*VaultSecret
-}
-
-type VaultSecret struct {
-	Name    string
-	Type    string
-	Path    string
-	Field   string
-	Version string
-}
-
-// names to assign to access attributes
-const (
-	ROLE_ID      = "roleID"
-	SECRET_ID    = "secretID"
-	TOKEN        = "token"
-	APPROLE_AUTH = "approle"
-	TOKEN_AUTH   = "token"
-	KV_V1        = "kv_v1"
-	KV_V2        = "kv_v2"
-)
-
-var masterAddress string
-var vaultClients map[string]*api.Client
-var InstanceAddresses map[string]bool
-var invalidInstances []string
-
-// Called once with toplevel/instance
-// Creates global map of all vault clients defined in a-i
-// This allows reconciliation of multiple vault instances
-func InitClients(instanceCreds map[string]AuthBundle, threadPoolSize int) {
-	InstanceAddresses = make(map[string]bool)
-	vaultClients = make(map[string]*api.Client)
-	configureMaster()
-
-	bwg := utils.NewBoundedWaitGroup(threadPoolSize)
-	var mutex = &sync.Mutex{}
-	// read access credentials for other vault instances and configure clients
-	for addr, bundle := range instanceCreds {
-		// client already configured separately for master
-		if addr != masterAddress {
-			bwg.Add(1)
-			go createClient(addr, bundle, &bwg, mutex)
-		}
-	}
-	bwg.Wait()
-
-	// set global for reference by toplevels to determine instances for reconciliation
-	for address := range vaultClients {
-		InstanceAddresses[address] = true
-	}
-}
-
-// goroutine support function for InitClients()
-// initializes one vault client
-func createClient(addr string, bundle AuthBundle, bwg *utils.BoundedWaitGroup, mutex *sync.Mutex) {
-	defer bwg.Done()
-
-	accessCreds := make(map[string]string)
-	for _, cred := range bundle.VaultSecrets {
-		// masterAddress hard-coded because all "child" vault access credentials must be pulled from master
-		processedCred, err := GetVaultSecretField(masterAddress, cred.Path, cred.Field, bundle.SecretEngine)
-		if err != nil {
-			log.WithError(err).Fatal()
-		}
-		accessCreds[cred.Name] = processedCred
-	}
-
-	// Init new client
-	config := api.DefaultConfig()
-	config.Address = addr
-	client, err := api.NewClient(config)
-	if err != nil {
-		log.WithError(err)
-		fmt.Println(fmt.Sprintf("Failed to initialize Vault client for %s", addr))
-		fmt.Println(fmt.Sprintf("SKIPPING ALL RECONCILIATION FOR: %s\n", addr))
-		return // skip entire reconcilation for this instance
-	}
-
-	// at minimum, one element will exist in secrets regardless of type
-	// type is same across all VaultSecrets associated with a particular instance address
-	var token string
-	switch bundle.VaultSecrets[0].Type {
-	case APPROLE_AUTH:
-		t, err := client.Logical().Write("auth/approle/login", map[string]interface{}{
-			"role_id":   accessCreds[ROLE_ID],
-			"secret_id": accessCreds[SECRET_ID],
-		})
-		if err != nil {
-			log.WithError(err)
-			fmt.Println(fmt.Sprintf("[Vault Client] failed to login to %s with AppRole credentials", addr))
-			fmt.Println(fmt.Sprintf("SKIPPING ALL RECONCILIATION FOR: %s\n", addr))
-			return // skip entire reconcilation for this instance
-		}
-		token = t.Auth.ClientToken
-	case TOKEN_AUTH:
-		token = accessCreds[TOKEN]
-	}
-
-	// add new address/client pair to global
-	mutex.Lock()
-	defer mutex.Unlock()
-	client.SetToken(token)
-
-	// test client
-	_, err = client.Sys().ListAuth()
-	if err != nil {
-		log.WithError(err)
-		fmt.Println(fmt.Sprintf("[Vault Client] failed to login to %s", addr))
-		fmt.Println(fmt.Sprintf("SKIPPING ALL RECONCILIATION FOR: %s\n", addr))
-		return
-	}
-
-	vaultClients[addr] = client
-}
 
 // attempts to read/proccess a single access credential for a particular vault instance
 func GetVaultSecretField(instanceAddr, path, field, engineVersion string) (string, error) {
@@ -153,67 +33,12 @@ func GetVaultSecretField(instanceAddr, path, field, engineVersion string) (strin
 	return secret[field].(string), nil
 }
 
-// configureMaster initializes vault client for the master instance
-// This is the only client configured using environment variables
-// env vars: VAULT_ADDR, VAULT_AUTHTYPE, VAULT_ROLE_ID, VAULT_SECRET_ID, VAULT_TOKEN
-func configureMaster() {
-	masterVaultCFG := api.DefaultConfig()
-	masterVaultCFG.Address = mustGetenv("VAULT_ADDR")
-
-	client, err := api.NewClient(masterVaultCFG)
-	if err != nil {
-		log.WithError(err).Fatal("failed to initialize master Vault client")
-	}
-
-	var clientToken string
-	switch authType := defaultGetenv("VAULT_AUTHTYPE", "approle"); strings.ToLower(authType) {
-	case APPROLE_AUTH:
-		roleID := mustGetenv("VAULT_ROLE_ID")
-		secretID := mustGetenv("VAULT_SECRET_ID")
-
-		secret, err := client.Logical().Write("auth/approle/login", map[string]interface{}{
-			"role_id":   roleID,
-			"secret_id": secretID,
-		})
-		if err != nil {
-			log.WithError(err).Fatal("[Vault Client] failed to login to master Vault with AppRole")
-		}
-		clientToken = secret.Auth.ClientToken
-	case TOKEN_AUTH:
-		clientToken = mustGetenv("VAULT_TOKEN")
-	default:
-		log.WithField("authType", authType).Fatal("[Vault Client] unsupported auth type")
-	}
-
-	client.SetToken(clientToken)
-	vaultClients[masterVaultCFG.Address] = client
-	masterAddress = masterVaultCFG.Address
-}
-
 // returns the vault client associated with instance address
 func getClient(instanceAddr string) *api.Client {
 	if vaultClients[instanceAddr] == nil {
 		log.Fatalf("[Vault Client] client does not exist for address: %s", instanceAddr)
 	}
 	return vaultClients[instanceAddr]
-}
-
-// AddInvalild is called by toplevel packages when an error is encountered while reconciling
-// The invalid instance is appended to a global and then processed within RemoveInstanceFromReconcile
-func AddInvalid(instanceAddr string) {
-	invalidInstances = append(invalidInstances, instanceAddr)
-}
-
-// Removes an instance from the global slice utilized by toplevels to target instances for reconciliation
-func RemoveInstanceFromReconciliation() {
-	for _, invalid := range invalidInstances {
-		if _, exists := InstanceAddresses[invalid]; exists {
-			delete(InstanceAddresses, invalid)
-			fmt.Println(fmt.Sprintf("SKIPPING REMAINING RECONCILIATION FOR %s", invalid))
-		}
-	}
-	// clear invalid
-	invalidInstances = nil
 }
 
 // return proper secret path format based upon kv version
